% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict_Envs.R
\name{Cross_forecast}
\alias{Cross_forecast}
\title{Cross_forecast}
\usage{
Cross_forecast(
  geno,
  pheno_train,
  env_info,
  predict_env,
  reshuffle,
  depend = "Marker",
  predict.method = "GBLUP"
)
}
\arguments{
\item{predict.method}{}
}
\value{
predict value
}
\description{
Cross_forecast
}
