% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06.Exhaustive_searfch_plotter.R
\name{Eh_plot}
\alias{Eh_plot}
\title{Eh_plot}
\usage{
Eh_plot(Correlation, dap_x, dap_y, p = NULL, Paras = NULL, Cor = NULL)
}
\arguments{
\item{Correlation}{A pop_cor data frame containing correlation values for different pairs.}

\item{dap_x}{Specifies the column for the x-axis variable.}

\item{dap_y}{Specifies the column for the y-axis variable.}

\item{p}{A parameter with a default value of 1.}

\item{Paras}{Vector containing the names of environmental parameters to be considered.}

\item{Cor}{Assigning variable column names.Default is Cor.}
}
\value{
A combined pop_cor plot.
}
\description{
Generate a combined plot for pairwise correlations using pop_cor data.
}
\examples{
env_trait<-env_trait_calculate(data=trait,trait="FTgdd",env="env_code")
#Run
\donttest{pop_cor<-Exhaustive_search(env_trait,PTT_PTR,80)
dap1<-pop_cor[,2][which(abs(pop_cor[, -c(1:4)]) ==
max(abs(pop_cor[, -c(1:4)])), arr.ind = TRUE)][1]
dap2<-pop_cor[,3][which(abs(pop_cor[, -c(1:4)]) ==
max(abs(pop_cor[, -c(1:4)])), arr.ind = TRUE)][1]
Eh_plot(pop_cor,dap1,dap2)}
}
