% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07.envMeanPara_plotter.R
\name{eMP_plot}
\alias{eMP_plot}
\title{eMP_plot}
\usage{
eMP_plot(
  data,
  size = NULL,
  shape = NULL,
  method = NULL,
  Paras = NULL,
  env_code = NULL,
  linewidth = NULL,
  alpha = NULL,
  linetype = NULL,
  linecolor = NULL
)
}
\arguments{
\item{data}{An environmental parameter mean data frame.}

\item{size}{Point size for scatter plot.}

\item{shape}{Point shape for scatter plot.}

\item{method}{Regression method for the smoothing line.}

\item{Paras}{Vector containing the names of environmental parameters
to be considered.}

\item{env_code}{Point color for the scatter plot by env_code.Defualt
is env_code.}

\item{linewidth}{Line width for the smoothing line.}

\item{alpha}{Alpha value for points in the scatter plot.}

\item{linetype}{Line type for the smoothing line.}

\item{linecolor}{Line color for the smoothing line.}
}
\value{
A plotter displaying the relationship between environmental
parameter means and the overall mean.
}
\description{
Generate a plot of environmental parameter means against the overall mean.
}
\examples{
\donttest{
env_trait<-env_trait_calculate(trait,"FTgdd","env_code")
envMeanPara<-EPM(data=env_trait,env_paras=PTT_PTR)
eMP_plot(envMeanPara,
Paras=c('DL', 'GDD', 'PTT', 'PTR', 'PTS'))}
}
