% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_env_info.R
\docType{data}
\name{env_info}
\alias{env_info}
\title{env_info Dataset}
\format{
A data frame with env_info
}
\source{
This dataset comes from the article.
}
\usage{
data(env_info)
}
\description{
This dataset contains information about env_info
}
\examples{
data(env_info)

}
\keyword{dataset}
