% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env_trait.R
\name{env_trait_calculate}
\alias{env_trait_calculate}
\title{env_trait_calculate}
\usage{
env_trait_calculate(data, trait, env, q25_75 = TRUE)
}
\arguments{
\item{data}{Input data}

\item{trait}{The trait need calculate}

\item{env}{All envs need calculate}

\item{q25_75}{Calculate q25 and q75 values. The defualt is TRUE.}
}
\value{
Trait mean values within different envs. if q25_75=T, also contains q25 and q75 values
}
\description{
Calculate traits mean in all envs
}
\examples{
env_trait<-env_trait(data=trait,trait="FTgdd",env="env_code")
}
