% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envMeanPara.R
\name{envMeanPara}
\alias{envMeanPara}
\title{envMeanPara}
\usage{
envMeanPara(data, env_paras, maxR_dap1 = NULL, maxR_dap2 = NULL, Paras)
}
\arguments{
\item{data}{env_trait data}

\item{env_paras}{env para info data}

\item{maxR_dap1}{as your like}

\item{maxR_dap2}{as your like}

\item{Paras}{All env para names}
}
\value{
A dat frame of envMeanPara
}
\description{
envMeanPara
}
\examples{
envMeanPara<-envMeanPara(data=env_trait, env_paras=PTT_PTR, maxR_dap1=18,maxR_dap2=43, Paras=Paras)
}
