% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07.envMeanPara.R
\name{EPM}
\alias{EPM}
\title{envMeanPara}
\usage{
EPM(data, env_paras, max_d1 = NULL, max_d2 = NULL, Paras = NULL)
}
\arguments{
\item{data}{Data frame containing trait mean values
within different environments.}

\item{env_paras}{Data frame containing environmental parameter information.}

\item{max_d1}{The starting day for calculating
environmental mean parameters.
Default is 18, which is used for the test data.}

\item{max_d2}{The ending day for calculating
environmental mean parameters.
Default is 43, which is used for the test data.}

\item{Paras}{Vector containing the names of environmental
parameters to be calculated.}
}
\value{
A data frame containing trait mean values and calculated
environmental mean parameters.
}
\description{
Calculate environmental mean parameters for a specific trait
within a specified time range.
This time range is calculated by the Exhaustive_search.
}
\examples{
\donttest{
env_trait<-env_trait_calculate(trait,"FTgdd","env_code")
envMeanPara<-EPM(data=env_trait,env_paras=PTT_PTR)
}

}
