% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl_calculate.R
\name{etl_calculate}
\alias{etl_calculate}
\title{etl_calculate}
\usage{
etl_calculate(data, trait, env, line, bycol)
}
\arguments{
\item{data}{env info data}

\item{trait}{trait you select}

\item{env}{your envs col}

\item{line}{your lines col}

\item{bycol}{order your env data by your select.}
}
\value{
your trait values within one env and one line
}
\description{
calculate your trait values within one env and one line
}
\examples{
etl<-etl_calculate(data=env_info,trait="FTgdd",env="env_code",bycol="lat")
}
