% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl_plotter.R
\name{etl_plotter}
\alias{etl_plotter}
\title{etl_plotter}
\usage{
etl_plotter(
  data,
  mean,
  env_cols = NULL,
  shape = NULL,
  size = NULL,
  linewidth = NULL,
  area_color = NULL,
  area_alpha = NULL
)
}
\arguments{
\item{data}{etl input}

\item{mean}{env_trait}

\item{env_cols}{env colors gradient}

\item{shape}{point shape, like ggplot2}

\item{size}{point size, like ggplot2}

\item{linewidth}{line width , like ggplot2}

\item{area_color}{q25-q75 area fill color}

\item{area_alpha}{q25-q75 area fill color alpha}
}
\value{
A plot of etl ,fill area start env-trait q25 to q75, red line is env-trait mean
}
\description{
etl_plotter
}
\examples{
etl_plotter(data=etl,mean=env_trait)
}
