% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06.Exhaustive_searfch_plotter.R
\name{Exhaustive_plotter}
\alias{Exhaustive_plotter}
\title{Exhaustive_plotter}
\usage{
Exhaustive_plotter(Correlation, dap_x, dap_y, p = NULL, Paras, Cor = NULL)
}
\arguments{
\item{Correlation}{A pop_cor data frame containing correlation values for different pairs.}

\item{dap_x}{Specifies the column for the x-axis variable.}

\item{dap_y}{Specifies the column for the y-axis variable.}

\item{p}{A parameter with a default value of 1.}

\item{Paras}{Vector containing the names of environmental parameters to be considered.}

\item{Cor}{Assigning variable column names.Default is Cor.}
}
\value{
A combined pop_cor plot.
}
\description{
Generate a combined plot for pairwise correlations using pop_cor data.
}
\examples{
\dontrun{Exhaustive_plotter(input=pop_cor,dap_x=122, dap_y=122,p=1,
Paras=c('DL', 'GDD', 'PTT', 'PTR', 'PTS'))}
}
