% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exhaustive_search.R
\name{Exhaustive_search}
\alias{Exhaustive_search}
\title{Exhaustive_search}
\usage{
Exhaustive_search(
  data,
  env_paras,
  searching_daps = NULL,
  p = NULL,
  dap_x = NULL,
  dap_y = NULL,
  LOO = NULL,
  Paras
)
}
\arguments{
\item{data}{env_trait data from env_trait_calculate}

\item{env_paras}{env_paras data}

\item{searching_daps}{depends on your file,set your para, default is 122 depends on example data}

\item{p}{defualt is 1}

\item{dap_x}{like searching_daps}

\item{dap_y}{like searching_daps}

\item{LOO}{defualt is 0}

\item{Paras}{your env}
}
\value{
pop cor matrix
}
\description{
Exhaustive_search
}
\examples{
pop_cor<-Exhaustive_search(data=env_trait, env_paras=PTT_PTR, searching_daps=122,
                                     p=1, dap_x=122,dap_y=122,LOO=0,Paras=Paras)
}
