% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict_lasso.R
\name{GE_CV}
\alias{GE_CV}
\title{Predict}
\usage{
GE_CV(
  pheno,
  geno,
  env,
  para,
  Para_Name,
  depend = NULL,
  fold = NULL,
  reshuffle = NULL,
  model,
  methods = NULL,
  ms1 = NULL,
  ms2 = NULL,
  ENalpha = NULL,
  GBM_params = NULL,
  nIter = NULL,
  burnIn = NULL,
  thin = NULL,
  SVM_cost = NULL,
  gamma = NULL,
  GBM_rounds = NULL
)
}
\arguments{
\item{pheno}{pheno info, from LbyE}

\item{geno}{geno info}

\item{env}{env info}

\item{para}{envMeanPara calculated by envMeanPara}

\item{Para_Name}{Your env para Name}

\item{depend}{Maker or Norm}

\item{fold}{Number of folds like rrblup}

\item{reshuffle}{Repeat the calculation for how many cycles}

\item{model}{use different model such rrBLUP LASSO RR EN}

\item{methods}{RM.G RM.GE RM.E}

\item{ms1}{remove the line for which the number of missing environment > ms1}

\item{ms2}{remove the line for which the number of missing environment > ms2}

\item{GBM_params}{give GBM hyper params to get more accuracy result}

\item{alpha}{used for EN model}
}
\value{
prediction result
}
\description{
Only RE.G based on Norm support LASSO RR EN RF and SVM model.Others are building....
Bayesian has been builded.
The adjust para are also building.such as ntree for RF
}
\examples{
out<-Predict(pheno=pheno,geno=geno,env=env_info,para=envMeanPara,Para_Name="PTT",
                       depend="maker",fold=2,reshuffle=5,methods="RM.G",
                       ms1=ms1,ms2=ms2)
}
