% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_geno.R
\docType{data}
\name{geno}
\alias{geno}
\title{geno Dataset}
\format{
A data frame with geno matrix
}
\source{
This dataset comes from the article.
}
\usage{
data(geno)
}
\description{
This dataset contains information about geno
}
\examples{
data(geno)

}
\keyword{dataset}
