% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2_rrBLUP.R
\name{h2_rrBLUP}
\alias{h2_rrBLUP}
\title{Title h2_rrBLUP caculate}
\usage{
h2_rrBLUP(trait, geno, envs)
}
\arguments{
\item{trait}{pheno_trait}

\item{geno}{geno_type_txt,0,1,2type}

\item{envs}{envs you want to calculate}
}
\value{
h2 in different envs
}
\description{
Title h2_rrBLUP caculate
}
\examples{
\dontrun{h2<-h2_rrBLUP(trait=trait,geno=geno,envs="env_code")}
}
