% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LbyE_corrplot.R
\name{LbyE_corrplot}
\alias{LbyE_corrplot}
\title{LbyE_corrplot}
\usage{
LbyE_corrplot(LbyE, cor_type = NULL, color = NULL)
}
\arguments{
\item{LbyE}{A data frame calculate by LbyE_calculate, the colnames are Lines, the rawnames are Envs.}

\item{cor_type}{Graphic format of the corr map.The defualt format is heatmap.}

\item{color}{Set the color gradient}
}
\value{
A correalation plot between lines and envs
}
\description{
Plot Lines and envs correlation
}
\examples{
\dontrun{LbyE_corrplot(LbyE=LbyE,cor_type="heatmap",color=c("blue","white","red"))}
}
