% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOOCV.R
\name{LOOCV}
\alias{LOOCV}
\title{LOOCV}
\usage{
LOOCV(
  data,
  input,
  env_paras,
  Para_Name,
  line,
  trait,
  maxR_dap1 = NULL,
  maxR_dap2 = NULL,
  p = NULL,
  filter = NULL
)
}
\arguments{
\item{data}{trait data}

\item{input}{env trait data}

\item{env_paras}{env para info}

\item{Para_Name}{Para name}

\item{line}{line name}

\item{trait}{trait name}

\item{maxR_dap1}{as you like depend on your data}

\item{maxR_dap2}{as you like depend on your data}

\item{p}{defualt=1}

\item{filter}{filter num}
}
\value{
LOOCV data frame
}
\description{
LOOCV
}
\examples{
LOOCV(maxR_dap1=18,maxR_dap2=43,data=trait,input=env_trait,env_paras=PTT_PTR,
               Para_Name="PTT",trait="FTgdd",line="line_code",p=1,filter=4)
}
