% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cross_vadiation.R
\name{MMGP}
\alias{MMGP}
\title{Cross Vadidation and Genomic Prediction Cross environments}
\usage{
MMGP(
  pheno,
  geno,
  env,
  para,
  Para_Name,
  depend = NULL,
  fold = NULL,
  reshuffle = NULL,
  model,
  methods = NULL,
  ms1 = NULL,
  ms2 = NULL,
  ENalpha = NULL,
  GBM_params = NULL,
  nIter = NULL,
  burnIn = NULL,
  thin = NULL,
  SVM_cost = NULL,
  gamma = NULL,
  GBM_rounds = NULL
)
}
\arguments{
\item{pheno}{Vector (n x j) of "phenotypes", i.e. observations or pre-processed, corrected values.}

\item{geno}{Matrix (n x m) of genotypes for the training population: n lines with m markers.
Genotypes should be coded -1, 0, 1. Missing data are not allowed.}

\item{env}{Data.frame (j x l) of "environmental information", i.e.}

\item{para}{Data.frame returned from envMeanPara function.}

\item{Para_Name}{The most relevant environmental silvers to the subject's phenotype,
obtained after stepwise correlation calculations, are referred to for more details:}

\item{depend}{The options for genomic breeding within different environment.The available options are:
1.Norm:
2.Marker}

\item{fold}{Number of folds for the cross-validation.
Smallest value recommended is fold = 2.}

\item{reshuffle}{Number of independent replicates for the cross-validation.
Smallest value recommended is reshuffle = 5.}

\item{model}{The options for genomic breeding cross vadiation methods. The available options are:
1.GBLUP: performs G-BLUP using a marker-based relationship matrix, implemented through BGLR R-library.
Equivalent to ridge regression (RR-BLUP) of marker effects.
2.RR: ridge regression, using package glmnet.
In theory, strictly equivalent to gblup.
3.LASSO: Least Absolute Shrinkage and Selection Operator is another penalized regression methods
which yield more shrinked estimates than RR.Run by glmnet library.
4.EN: Elastic Net (Zou and Hastie, 2005),
which is a weighted combination of RR and LASSO, using glmnet library
5.rrBLUP:
Several Bayesian methods, using the BGLR library:
1.BRR: Bayesian ridge regression: same as rr-blup, but bayesian resolution.
Induces homogeneous shrinkage of all markers effects towards zero with
Gaussian distribution (de los Campos et al, 2013)
2.BL: Bayesian LASSO: uses an exponential prior on marker variances priors,
leading to double exponential distribution of marker effects (Park & Casella 2008)
3.BA: uses a scaled-t prior distribution of marker effects. (Meuwissen et al 2001).
4.BB: Bayes B, uses a mixture of distribution with a point mass at zero
and with a slab of non-zero marker effects with a scaled-t distribution (Habier et al 2011).
5.BC: Bayes C same as Bayes B with a slab with Gaussian distribution.
A more detailed description of these methods can be found in Perez & de los Campos 2014 (http://genomics.cimmyt.org/BGLR-extdoc.pdf).
Four semi-parametric methods:
1.RKHS: reproductive kernel Hilbert space and multiple kernel MRKHS, using BGLR (Gianola and van Kaam 2008).
Based on genetic distance and a kernel function to regulate the distribution of marker effects.
This methods is claimed to be effective for detecting non additive effects.
2.RF: Random forest regression, using randomForest library (Breiman, 2001, Breiman and Cutler 2013).
This methods uses regression models on tree nodes which are rooted in bootstrapping data.
Supposed to be able to capture interactions between markers
3.SVM: support vector machine, run by e1071 library. For details,
see Chang, Chih-Chung and Lin, Chih-Jen: LIBSVM: a library for Support Vector Machines http://www.csie.ntu.edu.tw/~cjlin/libsvm
4.LightGBM: Light Gradient Boosting Machine, run by LightGBM library developed by Microsoft.
For details, see Jun Yan, Yuetong Xu, at al: LightGBM: accelerated genomically designed crop breeding through ensemble learning
and Mircosoft Mannual https://lightgbm.readthedocs.io/en/latest/R/index.html}

\item{methods}{RM.G RM.GE RM.E}

\item{ms1}{remove the line for which the number of missing environment > ms1}

\item{ms2}{remove the line for which the number of missing environment > ms2}

\item{ENalpha}{used for EN predict model,the value range from 0 to 1.}

\item{GBM_params}{a list of parameters. See The "Dataset Parameters"{https://lightgbm.readthedocs.io/en/latest/Parameters.html#dataset-parameters} section of the documentation}

\item{nIter}{Number of iterations for the Bayesian model.}

\item{burnIn}{Number of burn-in for the Bayesian model.}

\item{thin}{Number of thins for the Bayesian model.}

\item{SVM_cost}{cost of constraints violation (default: 10)—it is the 'C'-constant of the regularization term in the Lagrange formulation}

\item{gamma}{parameter needed for all kernels except linear (default: 0.001)}

\item{GBM_rounds}{number of training rounds Based on LightGBM model}
}
\value{
The class GP_CV returns a list containing:
The most relevant environmental silvers to the subject's phenotype,
obtained after stepwise correlation calculations, are referred to for more details:
}
\description{
Cross Vadidation and Genomic Prediction Cross environments
}
\examples{
\dontrun{
out<-MMGP(pheno=pheno,geno=geno,env=env_info,para=envMeanPara,Para_Name="PTT",
                       depend="maker",fold=2,reshuffle=5,methods="RM.G",
                       ms1=ms1,ms2=ms2)
}
}
