% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09.Predict_Envs.R
\name{MMPrdM}
\alias{MMPrdM}
\title{MMPrdM}
\usage{
MMPrdM(
  pheno,
  geno,
  env,
  para,
  Para_Name,
  model,
  depend = NULL,
  reshuffle = NULL,
  methods = NULL,
  ENalpha = NULL,
  SVM_cost = NULL,
  gamma = NULL,
  kernel = NULL,
  Envs = NULL,
  line_code = NULL,
  nIter = NULL,
  burnIn = NULL,
  thin = NULL,
  GBM_params = NULL,
  GBM_rounds = NULL
)
}
\arguments{
\item{pheno}{Vector (n x j) of "phenotypes", i.e. observations or pre-processed, corrected values.}

\item{geno}{Matrix (n x m) of genotypes for the training population:
n lines with m markers.
Genotypes should be coded -1, 0, 1.
Missing data are not allowed.}

\item{env}{Data.frame (j x l) of "environmental information", i.e.}

\item{para}{Data.frame returned from envMeanPara function.}

\item{Para_Name}{The most relevant environmental silvers to the subject's phenotype,
obtained after stepwise correlation calculations, are referred to for more details:}

\item{model}{The options for genomic breeding cross vadiation methods. The available options are:\if{html}{\out{<br>}}
1.GBLUP: performs G-BLUP using a marker-based relationship matrix, implemented through BGLR R-library.\if{html}{\out{<br>}}
Equivalent to ridge regression (RR-BLUP) of marker effects.\if{html}{\out{<br>}}
2.RR: ridge regression, using package glmnet.
In theory, strictly equivalent to gblup.\if{html}{\out{<br>}}
3.LASSO: Least Absolute Shrinkage and Selection Operator is another penalized regression methods\if{html}{\out{<br>}}
which yield more shrinked estimates than RR.Run by glmnet library.\if{html}{\out{<br>}}
4.EN: Elastic Net (Zou and Hastie, 2005),
which is a weighted combination of RR and LASSO, using glmnet library\if{html}{\out{<br>}}
5.rrBLUP:the RR-BLUP mixed model(Endelman ,2011). \if{html}{\out{<br>}}
One application is to estimate marker effects by ridge regression;\if{html}{\out{<br>}}
alternatively, BLUPs can be calculated based on an \if{html}{\out{<br>}}
additive relationship matrix or a Gaussian kernel.\if{html}{\out{<br>}}
Several Bayesian methods, using the BGLR library:\if{html}{\out{<br>}}
1.BRR: Bayesian ridge regression: same as rr-blup, but bayesian resolution.\if{html}{\out{<br>}}
Induces homogeneous shrinkage of all markers effects towards zero with\if{html}{\out{<br>}}
Gaussian distribution (de los Campos et al, 2013)\if{html}{\out{<br>}}
2.BL: Bayesian LASSO: uses an exponential prior on marker variances priors,\if{html}{\out{<br>}}
leading to double exponential distribution of marker effects (Park & Casella 2008)\if{html}{\out{<br>}}
3.BA: uses a scaled-t prior distribution of marker effects. (Meuwissen et al 2001).\if{html}{\out{<br>}}
4.BB: Bayes B, uses a mixture of distribution with a point mass at zero\if{html}{\out{<br>}}
and with a slab of non-zero marker effects with a scaled-t distribution (Habier et al 2011).\if{html}{\out{<br>}}
5.BC: Bayes C same as Bayes B with a slab with Gaussian distribution.\if{html}{\out{<br>}}
A more detailed description of these methods can be found in \if{html}{\out{<br>}}
Perez & de los Campos 2014 (http://genomics.cimmyt.org/BGLR-extdoc.pdf).\if{html}{\out{<br>}}
Four semi-parametric methods:\if{html}{\out{<br>}}
1.RKHS: reproductive kernel Hilbert space and multiple kernel MRKHS, using BGLR (Gianola and van Kaam 2008).\if{html}{\out{<br>}}
Based on genetic distance and a kernel function to regulate the distribution of marker effects.\if{html}{\out{<br>}}
This methods is claimed to be effective for detecting non additive effects.\if{html}{\out{<br>}}
2.RF: Random forest regression, using randomForest library (Breiman, 2001, Breiman and Cutler 2013).\if{html}{\out{<br>}}
This methods uses regression models on tree nodes which are rooted in bootstrapping data.\if{html}{\out{<br>}}
Supposed to be able to capture interactions between markers\if{html}{\out{<br>}}
3.SVM: support vector machine, run by e1071 library. For details,\if{html}{\out{<br>}}
see Chang, Chih-Chung and Lin, Chih-Jen: \if{html}{\out{<br>}}
LIBSVM: a library for Support Vector Machines http://www.csie.ntu.edu.tw/~cjlin/libsvm\if{html}{\out{<br>}}
4.LightGBM: Light Gradient Boosting Machine, run by LightGBM library developed by Microsoft.\if{html}{\out{<br>}}
For details, see Jun Yan, Yuetong Xu, at al: LightGBM: \if{html}{\out{<br>}}
accelerated genomically designed crop breeding through ensemble learning\if{html}{\out{<br>}}}

\item{depend}{The options for genomic breeding within different environment.The available options are:
1.Reaction Norm:
2.PEI:}

\item{reshuffle}{Number of independent replicates for the Predict.
Smallest value recommended is reshuffle = 3.}

\item{methods}{RM.G RM.GE RM.E}

\item{ENalpha}{used for EN predict model,the value range from 0 to 1.}

\item{SVM_cost}{Cost of constraints violation for SVM (default: 10).}

\item{gamma}{Parameter needed for all kernels except linear (default: 0.001).}

\item{kernel}{the kernel used in training and predicting.
You might consider changing some of the following parameters,
depending on the kernel type.
linear, polynomial, radial basis, sigmoid.}

\item{Envs}{Assign variable column names of Envs. Default is Envs.}

\item{line_code}{Assign variable column names of line_code Default is line_Code.}

\item{nIter}{Number of iterations for the Bayesian model.}

\item{burnIn}{Number of burn-in for the Bayesian model.}

\item{thin}{Number of thins for the Bayesian model.}

\item{GBM_params}{A list of parameters for LightGBM.
See LightGBM documentation for details.}

\item{GBM_rounds}{Number of training rounds for LightGBM.}
}
\value{
The class MMPrdM returns a list containing:
The Predicted Values of each line in Multiple Environments.
}
\description{
MMPrdM (Multi-environment Multi-marker Prediction Model) is a function
for genomic prediction that includes cross-environment prediction.
It leverages genetic and environmental information to predict phenotype.
The function supports various genomic prediction methods and models,
providing flexibility in choosing the most suitable approach
based on the data characteristics and user preferences.
}
\examples{
\dontrun{
out<-MMPrdM(pheno=pheno,geno=geno,env=env_info,para=envMeanPara)
}
}
