% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04.mse_plotter.R
\name{mse_plot}
\alias{mse_plot}
\title{mse_plot}
\usage{
mse_plot(data,point_size=NULL,text_size=NULL,point_shape=NULL)
}
\arguments{
\item{data}{Dataframe containing MSE data analyzed by line_trait_mean.}

\item{point_size}{Size of points in the plot.}

\item{text_size}{Size of text labels in the plot.}

\item{point_shape}{Shape of the points in the plot.}
}
\value{
A plot illustrating Mean Squared Errors (MSE) data.
}
\description{
Generate a plot for Mean Squared Errors (MSE) data analyzed by line_trait_mean.
}
\examples{
\donttest{
LbyE<-LbyE_calculate(trait,"FTgdd","env_code","line_code")
env_trait<-env_trait_calculate(trait,"FTgdd","env_code")
#Run
result<-line_trait_mean(trait,LbyE,env_trait,"FTgdd",row=2)

mse_plot(result[[1]])}
}
