% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict.R
\name{Predict}
\alias{Predict}
\title{Predict}
\usage{
Predict(
  pheno,
  geno,
  env,
  para,
  Para_Name,
  depend = NULL,
  fold = NULL,
  reshuffle = NULL,
  methods = NULL,
  ms1 = NULL,
  ms2 = NULL
)
}
\arguments{
\item{pheno}{pheno info, from LbyE}

\item{geno}{geno info}

\item{env}{env info}

\item{para}{envMeanPara calculated by envMeanPara}

\item{Para_Name}{Your env para Name}

\item{depend}{Maker or Norm}

\item{fold}{Number of folds like rrblup}

\item{reshuffle}{Repeat the calculation for how many cycles}

\item{methods}{RM.G RM.GE RM.E}

\item{ms2}{remove the line for which the number of missing environment > ms2}

\item{line_missing}{remove the environment for which the number of missing line > line_missing}
}
\value{
prediction result
}
\description{
Predict
}
\examples{
out<-Predict(pheno=pheno,geno=geno,env=env_info,para=envMeanPara,Para_Name="PTT",
                       depend="maker",fold=2,reshuffle=5,methods="RM.G",
                       ms1=ms1,ms2=ms2)
}
