% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_PTT_PTR.R
\docType{data}
\name{PTT_PTR}
\alias{PTT_PTR}
\title{PTT_PTR Dataset}
\format{
A data frame with env parameters
}
\source{
This dataset comes from the article.
}
\usage{
data(PTT_PTR)
}
\description{
This dataset contains information about PTT_PTR.
}
\examples{
data(PTT_PTR)

}
\keyword{dataset}
