% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Regression.R
\name{Reg}
\alias{Reg}
\title{Regression}
\usage{
Reg(LbyE, env_trait, filter_num = NULL)
}
\arguments{
\item{LbyE}{LbyE}

\item{env_trait}{env trait info}

\item{filter_num}{filter num}
}
\value{
Regression calculate result
}
\description{
Regression
}
\examples{
Reg(LbyE=LbyE,env_trait=env_trait)
}
