% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Slope_Intercept.R
\name{Slope_Intercept}
\alias{Slope_Intercept}
\title{Slope_Intercept}
\usage{
Slope_Intercept(
  maxR_dap1,
  maxR_dap2,
  Para_Name,
  data,
  line,
  trait,
  input,
  env_paras,
  filter = NULL,
  rounds = NULL
)
}
\arguments{
\item{maxR_dap1}{as your like}

\item{maxR_dap2}{as your like}

\item{Para_Name}{env para name}

\item{data}{trait data}

\item{line}{line code row}

\item{trait}{trait row}

\item{input}{env trait data}

\item{env_paras}{env para data}

\item{filter}{filter nums}

\item{rounds}{repeat nums}
}
\value{
Slope_Intercept dataframe
}
\description{
Slope_Intercept
}
\examples{
\dontrun{Slope_Intercept<-Slope_Intercept(data=trait,input=env_trait, env_paras=PTT_PTR,
                                 Para_Name="PTT",line="line_code",trait="FTgdd",filter=5,
                                 maxR_dap1=18, maxR_dap2=43,rounds=4)}
}
