% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00.data_trait.R
\docType{data}
\name{trait}
\alias{trait}
\title{trait Dataset}
\format{
A data frame representing trait parameters.
\describe{
\item{line_code}{Line code assigned to each line.}
\item{env_code}{Environmental code assigned to each environment.}
\item{env_note}{Environmental note assigned to each environment.}
\item{FTdap}{Name of the phenotypic trait measured, related to FT.}
\item{FTgdd}{Name of the phenotypic trait measured, related to FT.}
\item{pop_code}{Other columns may represent additional trait-related information.}
}
}
\source{
This dataset originates from the article and provides phenotypic trait data for analysis.
}
\usage{
data(trait)
}
\description{
This dataset contains information about phenotypic traits for various lines and environments.
}
\examples{
data(trait)

}
\keyword{dataset}
