% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict_Envs.R
\name{Cross_forecast}
\alias{Cross_forecast}
\title{Cross_forecast}
\usage{
Cross_forecast(
  geno,
  pheno_train,
  env_info,
  predict_env,
  reshuffle,
  depend,
  predict.method
)
}
\arguments{
\item{geno}{geno martrix}

\item{pheno_train}{pheno data frame used for training}

\item{env_info}{env info}

\item{predict_env}{env need predict}

\item{reshuffle}{times for duplication}

\item{depend}{use which depend}

\item{predict.method}{use which predict model, you can see this in MMGP}
}
\value{
predict value
}
\description{
Cross_forecast
}
\examples{
\dontrun{
Cross_forecast(geno,pheno_train,env_info,predict_env,reshuffle,depend,predict.method)
}
}
