% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00.data_env_info.R
\docType{data}
\name{env_info}
\alias{env_info}
\title{env_info Dataset}
\format{
A data frame with the following columns:
\describe{
\item{env_code}{Environmental code assigned to each environment.}
\item{env_notes}{Environmental note assigned to each environment.}
\item{lat}{Latitude of the environment.}
\item{lon}{Longitude of the environment.}
\item{Location}{Location of the environment.}
\item{PlantingDate}{Planting Date of lines.}
\item{TrialYear}{Planting Year of lines.}
}
}
\source{
This dataset originates from the article and provides essential information for environmental analysis.
}
\usage{
data(env_info)
}
\description{
This dataset contains information about environmental factors for various environments.
}
\examples{
data(env_info)

}
\keyword{dataset}
