% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.env_trait.R
\name{env_trait_calculate}
\alias{env_trait_calculate}
\title{env_trait_calculate}
\usage{
env_trait_calculate(data, trait, env, q25_75 = TRUE)
}
\arguments{
\item{data}{Input data frame containing trait values and
environmental information.}

\item{trait}{The name of the trait for which mean and optional
quantiles will be calculated.}

\item{env}{The column name in the data frame representing
different environments.}

\item{q25_75}{Logical, indicating whether to calculate the 25th and
75th quantiles. Default is TRUE}
}
\value{
Data frame containing the mean trait values for each environment.
If q25_75 is TRUE, it also includes columns for the 25th(q25), 75th(q75)
quantiles, as well as the number of observations (n_obs).
}
\description{
Calculate the mean and optional quantiles for a trait across
different environments.
}
\examples{
env_trait<-env_trait_calculate(trait,"FTgdd","env_code")

# Access the result, e.g., mean trait values
env_trait$mean

# Access the result, e.g., 25th quantile values
env_trait$q25

# Access the result, e.g., 75th quantile values
env_trait$q75

# Access the result, e.g., number of observations
env_trait$n_obs

}
