% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envMeanPara_plotter.R
\name{envMeanPara_plotter}
\alias{envMeanPara_plotter}
\title{envMeanPara_plotter}
\usage{
envMeanPara_plotter(
  envMeanPara,
  size = NULL,
  shape = NULL,
  method = NULL,
  linewidth = NULL,
  alpha = NULL,
  linetype = NULL,
  linecolor = NULL
)
}
\arguments{
\item{envMeanPara}{envMeanPara}

\item{size}{point size}

\item{shape}{point shape}

\item{method}{regression method}

\item{linewidth}{regression line width}

\item{alpha}{point alpha}

\item{linetype}{regression line type}

\item{linecolor}{regression line color}
}
\value{
A envparamean plotter
}
\description{
envMeanPara_plotter
}
\examples{
envMeanPara_plotter(envMeanPara)
}
