% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05.etl_plotter.R
\name{etl_plotter}
\alias{etl_plotter}
\title{etl_plotter}
\usage{
etl_plotter(
  data,
  trait,
  env_cols = NULL,
  shape = NULL,
  size = NULL,
  linewidth = NULL,
  area_color = NULL,
  area_alpha = NULL
)
}
\arguments{
\item{data}{An etl input data frame.}

\item{trait}{An env_trait data frame containing mean, q25, and q75 values.}

\item{env_cols}{Environmental colors gradient; if NULL,
it generates colors using rainbow_hcl.}

\item{shape}{Point shape, similar to ggplot2.}

\item{size}{Point size, similar to ggplot2.}

\item{linewidth}{Line width, similar to ggplot2.}

\item{area_color}{Fill color for the area between q25 and q75.}

\item{area_alpha}{Fill color alpha for the area between q25 and q75.}
}
\value{
A plot of etl, filling the area between env-trait q25 to q75,
with a red line representing the env-trait mean.
}
\description{
Generate a plot of environmental trait lines with mean, q25, and q75 areas.
}
\examples{
env_trait<-env_trait_calculate(trait,"FTgdd","env_code")
LbyE<-LbyE_calculate(trait,"FTgdd","env_code","line_code")
etl<-LbyE_Reshape(env_trait,LbyE,"env_code")
#Plot
etl_plotter(data=etl,trait=env_trait)


}
