% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06.Exhaustive_searfch_plotter.R
\name{Exhaustive_plotter}
\alias{Exhaustive_plotter}
\title{Exhaustive_plotter}
\usage{
Exhaustive_plotter <- function(Correlation,Paras,...)
}
\arguments{
\item{Correlation}{A pop_cor data frame containing correlation values for different pairs.}

\item{dap_x}{Specifies the column for the x-axis variable.}

\item{dap_y}{Specifies the column for the y-axis variable.}

\item{p}{A parameter with a default value of 1.}

\item{Paras}{Vector containing the names of environmental parameters to be considered.}

\item{Cor}{Assigning variable column names.Default is Cor.}
}
\value{
A combined pop_cor plot.
}
\description{
Generate a combined plot for pairwise correlations using pop_cor data.
}
\examples{
\dontrun{Exhaustive_plotter(Correlation,dap_x,dap_y,Paras)}

}
