% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exhaustive_plotter.R
\name{Exhaustive_plotter}
\alias{Exhaustive_plotter}
\title{envMeanPara_plotter}
\usage{
Exhaustive_plotter(data, dap_x, dap_y, p)
}
\arguments{
\item{data}{pop_cor file}

\item{dap_x}{depend on your file}

\item{dap_y}{depend on your file}

\item{p}{defualt p=1}
}
\value{
combined pop_cor plot
}
\description{
envMeanPara_plotter
}
\examples{
Exhaustive_plotter(input=pop_cor,paras="PTS")
}
