% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06.Exhaustive_search.R
\name{Exhaustive_search}
\alias{Exhaustive_search}
\title{Exhaustive_search}
\usage{
Exhaustive_search(
  data,
  env_paras,
  searching_daps,
  p = NULL,
  dap_x = NULL,
  dap_y = NULL,
  LOO = NULL,
  Paras = NULL
)
}
\arguments{
\item{data}{Data frame containing environmental trait data.}

\item{env_paras}{Data frame containing environmental parameter information.}

\item{searching_daps}{The total number of days to search for,
typically set based on your data. Default is 122.}

\item{p}{The parameter for controlling the shape of the search window.
Default is 1.}

\item{dap_x}{The number of days for the x-axis of the search window.
Default is the same as searching_daps.}

\item{dap_y}{The number of days for the y-axis of the search window.
Default is the same as searching_daps.}

\item{LOO}{Leave-One-Out cross-validation flag. If LOO is 1,}

\item{Paras}{Vector containing the names of environmental parameters
to be considered.}
}
\value{
A population correlation matrix based on the exhaustive search.
}
\description{
Perform exhaustive search for population
correlation matrix based on environmental parameters.
}
\examples{
#Input from function : env_trait_calculate
env_trait<-env_trait_calculate(data=trait,trait="FTgdd",env="env_code")
#Run
\donttest{Exhaustive_search(env_trait,PTT_PTR,80)}

}
