% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00.data_geno.R
\docType{data}
\name{geno}
\alias{geno}
\title{geno Dataset}
\format{
An object of class \code{data.frame} with 237 rows and 1463 columns.
}
\source{
This dataset originates from the article and provides genetic data for analysis.
}
\usage{
data(geno)
}
\description{
This dataset contains genetic information for various lines.
}
\examples{
data(geno)

}
\keyword{dataset}
