% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LbyE_calculate.R
\name{LbyE_calculate}
\alias{LbyE_calculate}
\title{LbyE_calculate}
\usage{
LbyE_calculate(data, trait, env, line)
}
\arguments{
\item{data}{Input data}

\item{trait}{Trait you select}

\item{env}{All envs need calculate}

\item{line}{All Lines need calculate}
}
\value{
Lines values within Envs.
}
\description{
calculate Lines in all envs
}
\examples{
\dontrun{LbyE<-LbyE_calculate(data=data,trait="FTgdd",env="env_code",line="line_code")}
}
