% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02.LbyE_corrplot.R
\name{LbyE_corrplot}
\alias{LbyE_corrplot}
\title{LbyE_corrplot}
\usage{
LbyE_corrplot(LbyE, cor_type = NULL, color = NULL)
}
\arguments{
\item{LbyE}{A data frame calculated by LbyE_calculate. Column names
represent Lines, and row names represent Environments.}

\item{cor_type}{Graphic format of the correlation map. The default format
is "heatmap". Other option is "pie".}

\item{color}{Set the color gradient for the correlation map. It is a
vector of three colors representing low, mid, and high values.}
}
\value{
A correlation plot between lines and environments in
the specified format.
}
\description{
Plot the correlation between lines and environments.
}
\examples{
LbyE<-LbyE_calculate(trait,trait="FTgdd",env="env_code",line="line_code")
# Generate a heatmap correlation plot for Lines and Environments
LbyE_corrplot(LbyE,cor_type="heatmap",color=c("blue","white","red"))
# Generate a pie chart correlation plot for Lines and Environments
LbyE_corrplot(LbyE,cor_type="pie",color=c("blue","white","red"))
}
