% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_trait_mean.R
\name{line_trait_mean}
\alias{line_trait_mean}
\title{line_trait_mean}
\usage{
line_trait_mean(data, LbyE, mean, trait, row)
}
\arguments{
\item{data}{data input}

\item{LbyE}{LbyE input}

\item{mean}{env_trait input}

\item{trait}{mean rows in env_trait}

\item{row}{at least rows}
}
\value{
line_trait_mean
}
\description{
line_trait_mean
}
\examples{
\dontrun{line_trait_mean(LbyE=LbyE,mean=env_trait,row=2)}

}
