% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04.line_trait_mean.R
\name{line_trait_mean}
\alias{line_trait_mean}
\title{line_trait_mean}
\usage{
line_trait_mean(data, LbyE, mean, trait, row)
}
\arguments{
\item{data}{Input data frame containing trait values, line information,
and environmental information.}

\item{LbyE}{Data frame calculated by LbyE_calculate,
containing line values within different environments.}

\item{mean}{Data frame containing mean trait values
within different environments.
This data frame get from the function of env_mean_calculate.}

\item{trait}{The trait for which the residuals will be calculated.}

\item{row}{The minimum number of rows required for
fitting a linear model for each line.}
}
\value{
A list containing two elements:
\enumerate{
\item data: Data frame with columns env_code, errors (mean of residuals),
and additional columns from env_trait.
\item lm_residuals: Data frame containing the residuals squared for
each line and environment.
}
}
\description{
Calculate the mean of residuals for each line after fitting a linear model.
}
\examples{
LbyE<-LbyE_calculate(trait,trait="FTgdd",env="env_code",line="line_code")
env_trait<-env_trait_calculate(data=trait,trait="FTgdd",env="env_code")
#Run
result<-line_trait_mean(data=trait,LbyE=LbyE,
                       mean = env_trait,trait="FTgdd",row=2)
print(result[[1]])


}
