% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04.Mean_trait_plot.R
\name{Mean_trait_plot}
\alias{Mean_trait_plot}
\title{Mean_trait_plot}
\usage{
Mean_trait_plot(
  Reg,
  MSE,
  point_size = NULL,
  color = NULL,
  point_shape = NULL,
  linewidth = NULL,
  linetype = NULL
)
}
\arguments{
\item{Reg}{Regression result data frame.}

\item{MSE}{Mean Squared Error (MSE) result data frame.}

\item{point_size}{Size of points in the plot.}

\item{color}{color for the points and the dash area.}

\item{point_shape}{Shape of points in the plot.}

\item{linewidth}{Line width for the smoothing line.}

\item{linetype}{Line type for the smoothing line.}
}
\value{
A plot showing the relationship between mean trait values and
the environmental parameter mean, with MSE information.
}
\description{
Generate a plot displaying the relationship between the mean trait values
and the environmental parameter mean.
}
\examples{
\dontrun{Mean_trait_plot(Reg, MSE)}

}
