% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mean_trait_plot.R
\name{Mean_trait_plot}
\alias{Mean_trait_plot}
\title{Mean_trait_plot}
\usage{
Mean_trait_plot(
  Reg,
  MSE,
  point_size = NULL,
  point_shape = NULL,
  linewidth = NULL,
  linetype = NULL
)
}
\arguments{
\item{Reg}{Reg result}

\item{MSE}{MSE reuslt}

\item{point_size}{point_size}

\item{point_shape}{point_shape}

\item{linewidth}{linewidth}

\item{linetype}{linetype}
}
\value{
A MSE plot. x is Popmean ,y is MES. plot by ggplot2
}
\description{
Mean_trait_plot
}
\examples{
\dontrun{Mean_trait_plot(Reg,MSE)}
}
