% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08.Cross_vadiation.R
\name{MMGP}
\alias{MMGP}
\title{Cross Vadidation and Genomic Prediction Cross environments}
\usage{
MMGP(
  pheno,
  geno,
  env,
  para,
  Para_Name,
  model,
  depend = NULL,
  fold = NULL,
  reshuffle = NULL,
  methods = NULL,
  ENalpha = NULL,
  SVM_cost = NULL,
  gamma = NULL,
  kernel = NULL,
  nIter = NULL,
  burnIn = NULL,
  thin = NULL,
  GBM_params = NULL,
  GBM_rounds = NULL,
  Envs = NULL,
  Obs = NULL,
  line_code = NULL,
  ms1 = NULL,
  ms2 = NULL
)
}
\arguments{
\item{pheno}{Vector (n x j) of "phenotypes," i.e.,
observations or pre-processed, corrected values.}

\item{geno}{Matrix (n x m) of genotypes for the training population:
n lines with m markers.Genotypes should be coded -1, 0, 1.
Missing data are not allowed.}

\item{env}{Data.frame (j x l) of "environmental information,"
i.e., environmental conditions corresponding to phenotypes.}

\item{para}{Data.frame returned from envMeanPara function,
providing information about environmental means and parameters.}

\item{Para_Name}{The most relevant environmental covariates to the subject's phenotype,
obtained after stepwise correlation calculations.}

\item{model}{The options for genomic breeding cross vadiation methods.}

\item{depend}{The options for genomic breeding within different environment.
The available options are:
1.Norm(Reaction Norm Model)
2.PEI(Polygenic Environment Interaction Model)}

\item{fold}{Number of folds for cross-validation.
Smallest recommended value is fold = 2.}

\item{reshuffle}{Number of independent replicates for cross-validation.
Smallest recommended value is reshuffle = 5.\if{html}{\out{<br>}}}

\item{methods}{A vector specifying the prediction methods to be used.}

\item{ENalpha}{Used for Elastic Net prediction model,
with a value range from 0 to 1.}

\item{SVM_cost}{Cost of constraints violation for SVM (default: 10).}

\item{gamma}{Parameter needed for all kernels except linear (default: 0.001).}

\item{kernel}{the kernel used in training and predicting.
You might consider changing some of the following parameters,
depending on the kernel type.
linear, polynomial, radial basis, sigmoid.}

\item{nIter}{Number of iterations for the Bayesian model.}

\item{burnIn}{Number of burn-in iterations for the Bayesian model.}

\item{thin}{Thinning parameter for the Bayesian model.}

\item{GBM_params}{A list of parameters for LightGBM.
See LightGBM documentation for details.}

\item{GBM_rounds}{Number of training rounds for LightGBM.}

\item{Envs}{Assign variable column names of Envs. Default is Envs.}

\item{Obs}{Assign variable  names of Observed values. Default is Obs.}

\item{line_code}{Assign variable column names of line_code.Default is line_code.}

\item{ms1}{Remove the line for which the number of missing environments > ms1.}

\item{ms2}{Remove the line for which the number of missing environments > ms2.}
}
\value{
The class MMGP returns a list containing:
The Observed and Predicted Values of each line in Multiple Environments.
The Predictive accuracy in Multiple Environments.Calculated by Pearson Cor.
The Predictive accuracy for All Environments.(Cross R2)
}
\description{
MMGP (Multi-environment Multi-Genomic Prediction) is a function
for genomic prediction that incorporates cross-validation and
cross-environment prediction. It utilizes both
genetic and environmental information to predict phenotypic values.
The function supports various genomic prediction methods and models,
allowing users to choose the most suitable approach
based on data characteristics and preferences.
}
\examples{
\donttest{
env_trait<-env_trait_calculate(data=trait,trait="FTgdd",env="env_code")
LbyE<-LbyE_calculate(data=trait,trait="FTgdd",env="env_code",line="line_code")
pheno<-LbyE[which(as.character(LbyE$line_code)\%in\%
c("line_code",as.character(geno$line_code))),]

env_trait<-env_trait_calculate(trait,"FTgdd","env_code")
envMeanPara<-EPM(data=env_trait,env_paras=PTT_PTR)

out<-MMGP(pheno=pheno,geno=geno,env=env_info,
         para=envMeanPara,Para_Name="PTT",model="rrBLUP",
         depend="Norm",fold=2,reshuffle=5,methods="RM.G",
         ms1=2,ms2=2)
}
}
