% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mse_plotter.R
\name{mse_plotter}
\alias{mse_plotter}
\title{mse_plotter}
\usage{
mse_plotter(data, point_size = NULL, text_size = NULL, point_shape = NULL)
}
\arguments{
\item{data}{mse data analysised by line_trait_mean}

\item{point_size}{point size}

\item{text_size}{text size}

\item{point_shape}{point shape}
}
\value{
A ltm ploter
}
\description{
mse_plotter
}
\examples{
mse_plotter(ltm)
}
