% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03.Regression.R
\name{Reg}
\alias{Reg}
\title{Regression}
\usage{
Reg(LbyE, env_trait, filter_num = NULL)
}
\arguments{
\item{LbyE}{Data frame containing lines and their corresponding
trait values within environments.It is from the function of LbyE_calculate.}

\item{env_trait}{Data frame containing environmental means for
different traits within environments.It is from the function of enV_mean_calculate.}

\item{filter_num}{Minimum number of observations required to perform regression. Default is 4.}
}
\value{
A data frame containing regression results for each line.
}
\description{
Calculate regression results between traits and environmental means for each line.
}
\examples{
#Get Input
LbyE <- LbyE_calculate(data = trait, trait = "FTgdd", env = "env_code", line = "line_code")
env_trait <- env_trait_calculate(data = trait, trait = "FTgdd", env = "env_code", q25_75 = TRUE)
#Calculate
Regression<-Reg(LbyE = LbyE, env_trait = env_trait)
print(head(Regression))
}
