% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reg_plotter.R
\name{Reg_plotter}
\alias{Reg_plotter}
\title{Title}
\usage{
Reg_plotter(Reg = Reg, size = NULL, method = NULL, color = NULL, alpha = NULL)
}
\arguments{
\item{Reg}{input data}

\item{size}{piont size}

\item{method}{smooth method}

\item{color}{line and point color}

\item{alpha}{line and point alpha}
}
\value{
A regression plot between mean and trait
}
\description{
Title
}
\examples{
Reg_plotter(Reg=Reg)
}
