% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_trait.R
\docType{data}
\name{trait}
\alias{trait}
\title{trait Dataset}
\format{
A data frame with trait parameters
}
\source{
This dataset comes from the article.
}
\usage{
data(trait)
}
\description{
This dataset contains information about trait
}
\examples{
data(trait)

}
\keyword{dataset}
